fix: restore the `types` field in the `package.json` file
TypeScript may not use the `"types"` import condition.

For example, if the value of the `moduleResolution` option is `"node"`. In this case,
TypeScript uses the `types` field and the `"types"` import condition is not used.

So even if the `"types"` import condition is used, `package.json` still needs the `types` field
for fallback. Therefore, I have restored the `types` field
that was removed in f7240436632735fdab258c5fc8e1456c9e2a0db1.

Resolves #26